refactor: move binary base class to existing file
